Update site.pp to work with Apache+SSL
NOTE: every keystone CLI command (including the one run inside puppet-keystone
module) need to be run with --insecure option as of now. Also, the cert file and
the private key need to be generated and kept at a location, with user and group
set to "keystone". All these limitations will be removed in future.

The modified puppet-keystone module will be placed in a reasonable location
soon, and the Puppetfile will be updated in a later commit